rename txn_hash -> txn_snapshot_word_serializer
