+++ silkworm/db/snapshots/snapshot_test.cpp
@@ -31,7 +31,7 @@
 #include <silkworm/infra/common/log.hpp>
 #include <silkworm/infra/test_util/log.hpp>
 
-#include "txn_hash.hpp"
+#include "txn_snapshot_word_serializer.hpp"
 
 namespace silkworm::snapshots {
