+++ silkworm/db/snapshots/txn_index.cpp
@@ -17,7 +17,7 @@
 #include "txn_index.hpp"
 
 #include "snapshot.hpp"
-#include "txn_hash.hpp"
+#include "txn_snapshot_word_serializer.hpp"
 
 namespace silkworm::snapshots {
